rename "remove positioning" component
