+++ packages/app/src/runner/screenshot/AdjustRunnerStyleDuringScreenshot.vue
